refactor: remove old Fmgc global barrel and fix mistaken imports
