+++ fbw-a32nx/src/systems/extras-host/modules/flightplan_sync/FlightPlanAsoboSync.ts
@@ -14,13 +14,6 @@ import {
   SyncFlightPlanEvents,
 } from '@fmgc/flightplanning/sync/FlightPlanEvents';
 import {
-  A320FlightPlanPerformanceData,
-  FlightPlanIndex,
-  NavigationDatabase,
-  NavigationDatabaseBackend,
-  NavigationDatabaseService,
-} from '@fmgc/index';
-import {
   EventBus,
   FacilityType,
   FacilityLoader,
@@ -33,6 +26,10 @@ import {
 } from '@microsoft/msfs-sdk';
 import { ApproachType as MSApproachType } from '../../../../../../fbw-common/src/systems/navdata/client/backends/Msfs/FsTypes';
 import { FacilityCache } from '../../../../../../fbw-common/src/systems/navdata/client/backends/Msfs/FacilityCache';
+import { NavigationDatabaseService } from '@fmgc/flightplanning/NavigationDatabaseService';
+import { NavigationDatabase, NavigationDatabaseBackend } from '@fmgc/NavigationDatabase';
+import { A320FlightPlanPerformanceData } from '@fmgc/flightplanning/plans/performance/FlightPlanPerformanceData';
+import { FlightPlanIndex } from '@fmgc/flightplanning/FlightPlanManager';
 
 export class FlightPlanAsoboSync {
   private isReady = false;
@@ -66,6 +63,7 @@ export class FlightPlanAsoboSync {
   }
 
   init(): void {
+    // FIXME this should only ever be used within the FMGC
     NavigationDatabaseService.activeDatabase = new NavigationDatabase(NavigationDatabaseBackend.Msfs);
 
     const sub = this.bus.getSubscriber<
