+++ fbw-a32nx/src/systems/fmgc/src/navigation/Navigation.ts
@@ -4,14 +4,15 @@
 
 import { Arinc429Register, IlsNavaid, NdbNavaid, VhfNavaid, VhfNavaidType, Icao } from '@flybywiresim/fbw-sdk';
 
-import { EventBus, FlightPlanService } from '@fmgc/index';
 import { LandingSystemSelectionManager } from '@fmgc/navigation/LandingSystemSelectionManager';
 import { NavaidSelectionManager, VorSelectionReason } from '@fmgc/navigation/NavaidSelectionManager';
 import { NavaidTuner } from '@fmgc/navigation/NavaidTuner';
 import { NavigationProvider } from '@fmgc/navigation/NavigationProvider';
 import { NearbyFacilities } from '@fmgc/navigation/NearbyFacilities';
 import { RequiredPerformance } from '@fmgc/navigation/RequiredPerformance';
+import { EventBus } from '@microsoft/msfs-sdk';
 import { Coordinates } from 'msfs-geo';
+import { FlightPlanService } from '../flightplanning/FlightPlanService';
 
 export enum SelectedNavaidType {
   None,
