+++ fbw-a32nx/src/systems/instruments/src/MCDU/legacy_pages/A320_Neo_CDU_DirectToPage.ts
@@ -1,12 +1,12 @@
 // Copyright (c) 2020, 2022 FlyByWire Simulations
 // SPDX-License-Identifier: GPL-3.0
 
-import { WaypointEntryUtils } from '@fmgc/index';
 import { CDUFlightPlanPage } from './A320_Neo_CDU_FlightPlanPage';
 import { NXFictionalMessages, NXSystemMessages } from '../messages/NXSystemMessages';
 import { Keypad } from '../legacy/A320_Neo_CDU_Keypad';
 import { Fix } from '@flybywiresim/fbw-sdk';
 import { LegacyFmsPageInterface } from '../legacy/LegacyFmsPageInterface';
+import { WaypointEntryUtils } from '@fmgc/flightplanning/WaypointEntryUtils';
 
 // TODO this whole thing is thales layout...
