+++ fbw-a32nx/src/systems/instruments/src/MCDU/legacy_pages/A320_Neo_CDU_VerticalRevisionPage.ts
@@ -3,7 +3,6 @@
 // SPDX-License-Identifier: GPL-3.0
 
 import { WaypointConstraintType } from '@fmgc/flightplanning/data/constraint';
-import { FlightPlanIndex } from '@fmgc/index';
 import { FmgcFlightPhase } from '@shared/flightphase';
 import { Keypad } from '../legacy/A320_Neo_CDU_Keypad';
 import { CDUFlightPlanPage } from './A320_Neo_CDU_FlightPlanPage';
@@ -12,6 +11,7 @@ import { CDUWindPage } from './A320_Neo_CDU_WindPage';
 import { NXSystemMessages } from '../messages/NXSystemMessages';
 import { AltitudeDescriptor } from '@flybywiresim/fbw-sdk';
 import { LegacyFmsPageInterface } from '../legacy/LegacyFmsPageInterface';
+import { FlightPlanIndex } from '@fmgc/flightplanning/FlightPlanManager';
 
 export class CDUVerticalRevisionPage {
   /**
