+++ fbw-a32nx/src/systems/instruments/src/MCDU/legacy_pages/atsu/FansB/A320_Neo_CDU_ATC_UsualRequest.ts
@@ -1,10 +1,10 @@
 import { AtsuStatusCodes, CpdlcMessage, CpdlcMessagesDownlink, InputValidation } from '@datalink/common';
-import { WaypointEntryUtils } from '@fmgc/index';
 import { Keypad } from '../../../legacy/A320_Neo_CDU_Keypad';
 import { CDUAtcMenu } from '../A320_Neo_CDU_ATC_Menu';
 import { CDUAtcTextFansB } from '../FansB/A320_Neo_CDU_ATC_Text';
 import { NXSystemMessages } from '../../../messages/NXSystemMessages';
 import { LegacyAtsuPageInterface } from '../../../legacy/LegacyAtsuPageInterface';
+import { WaypointEntryUtils } from '@fmgc/flightplanning/WaypointEntryUtils';
 
 export class CDUAtcUsualRequestFansB {
   static CreateDataBlock() {
