+++ fbw-a380x/src/systems/instruments/src/MFD/pages/FMS/DATA/MfdFmsDataStatus.tsx
@@ -10,9 +10,9 @@ import { MfdSimvars } from 'instruments/src/MFD/shared/MFDSimvarPublisher';
 import { TopTabNavigator, TopTabNavigatorPage } from 'instruments/src/MsfsAvionicsCommon/UiWidgets/TopTabNavigator';
 import { Button } from 'instruments/src/MsfsAvionicsCommon/UiWidgets/Button';
 import { AirlineModifiableInformation } from '@shared/AirlineModifiableInformation';
-import { NavigationDatabaseService } from '@fmgc/index';
 import { DatabaseIdent } from '@flybywiresim/fbw-sdk';
 import { ConfirmationDialog } from '../../../../MsfsAvionicsCommon/UiWidgets/ConfirmationDialog';
+import { NavigationDatabaseService } from '@fmgc/flightplanning/NavigationDatabaseService';
 
 import './MfdFmsDataStatus.scss';
