+++ fbw-a380x/src/systems/instruments/src/MFD/pages/FMS/F-PLN/FplnRevisionsMenu.tsx
@@ -1,7 +1,7 @@
 import { TurnDirection, WaypointDescriptor } from '@flybywiresim/fbw-sdk';
 import { HoldType } from '@fmgc/flightplanning/data/flightplan';
+import { FlightPlanIndex } from '@fmgc/flightplanning/FlightPlanManager';
 import { SegmentClass } from '@fmgc/flightplanning/segments/SegmentClass';
-import { FlightPlanIndex } from '@fmgc/index';
 import { MfdFmsFpln } from 'instruments/src/MFD/pages/FMS/F-PLN/MfdFmsFpln';
 import { ContextMenuElement } from 'instruments/src/MsfsAvionicsCommon/UiWidgets/ContextMenu';
 import { BitFlags } from '@microsoft/msfs-sdk';
