+++ fbw-a380x/src/systems/systems-host/systems/BrakeToVacateDistanceUpdater.ts
@@ -10,9 +10,10 @@ import {
   pointDistance,
 } from '@flybywiresim/oanc';
 import { Arinc429Register, Arinc429SignStatusMatrix, MathUtils } from '@flybywiresim/fbw-sdk';
-import { NavigationDatabase, NavigationDatabaseBackend, NavigationDatabaseService } from '@fmgc/index';
 import { placeBearingDistance } from 'msfs-geo';
 import { Position } from '@turf/turf';
+import { NavigationDatabaseService } from '@fmgc/flightplanning/NavigationDatabaseService';
+import { NavigationDatabase, NavigationDatabaseBackend } from '@fmgc/NavigationDatabase';
 
 /**
  * Utility class for brake to vacate (BTV) functions on the A380
